Prevent backend startup if relationship columns are missing
An easy mistake to make is to define a new relationship type but
forget to add the full set of columns to the underlying relationship
table.  If you test as admin everything looks OK, but then things
break for non-admins.

Modify the startup to check that the tables look valid and refuse to
continue if columns are missing.

(cherry picked from commit 6e0418a7dc16a7147926a6ce8e746375a7fab4f7)