Add support for skipping relationship fetch selectively
Since The Beginning Of Time it's been possible to pass a
`:skip_relationships` boolean on a sequel_to_jsonmodel call in order
to avoid pulling back a bunch of relationship information that a
caller doesn't need.

This commit extends this functionality to allow specific relationships
to be skipped, rather than all of them.  In our case, we need this for
a record type that might link to many thousands of records: we don't
need that full list of records when querying, but we do need other
relationship fields.

(cherry picked from commit 3fee2b8e19e1e09b3e510128d2db74cdfbaffce0)