Find and fix zombie records
There's a bug in MySQL versions prior to 8.0 that can cause a newly created
record to receive the same AUTO_INCREMENT id as a previously deleted record.

Circumstances where this will happen are:

 * A record of some type is created and then deleted

 * MySQL is restarted

 * A new record of that same type is created

The bug is described here:

  https://bugs.mysql.com/bug.php?id=199

The short version: MySQL versions < 8.0 keep track of the next auto increment
ID to assign for each table *in memory only*.  When the DB is restarted, the
in-memory counter is recomputed as MAX(id) + 1 and the sequence starts from
there.  Since any number of records following MAX(id) could have been deleted
prior to the restart (and still represented in the `deleted_records` table as
Tombstones), this is a problem.

This commit adds a scheduled job which scans the deleted record table
for any record whose ID is still in use in one of the live record
tables.  When it finds one, it:

 * Removes the tombstone

 * Marks the record for reindex

 * Allows plugin code to do its own cleanup by invoking registered callbacks
 * (see below)

All of this can go away once everyone moves off MySQL 5.x.

(cherry picked from commit 0df7e84de827d90360fd155d9896de23a0bf887b)