Add a new jsonstore for storing indexed JSON blobs
This significantly reduces the size of our Solr indexes, by storing
JSON records in an offside SQLite file.  Records are packed together
and compressed using zstd.  At search time, the appropriate version of
the JSON record is pulled from the JSON store and included in the
results, same as it ever was.

See the top of jsonstore.rb for more information about how it all works.