Optimise nested resolving
Profiler shows that quite a lot of time was spent in
`find_matching_refs`, which is responsible for finding the

  {'ref' => ...}

elements of interest in a heavily nested JSON tree.

The previous approach worked a record at a time, then a resolve
property at a time, to find the relevant refs and pluck them out.  As
a result, it ended up doing work proportional to NUMBER_OF_RECORDS *
NUMBER_OF_RESOLVED_PROPERTIES.

The new implementation handles all resolve properties at once.  It
walks the entire set of records and, when it finds a ref, checks
whether the property it was found under is one that is marked to be
resolved.  This means we just walk the JSON once, and aren't
particularly sensitive to the number of properties being resolved.

(cherry picked from commit ca857726933096d9fe961ddc311b5d022c6886d3)