PERFORMANCE Avoid indexing large changes twice
The periodic indexer always looks back 30 seconds prior to the time it
last checked for updates, and this means it can index the same set of
records more than once.

This often doesn't matter (and it's better than missing records), but
when there's a large number of updates, it's costly to index them
twice and holds up other records that haven't even been indexed once
yet.

The scenario where things go bad is like this:

  * Large number of updates recorded with system_mtime=T100

  * Indexer runs at T105 and picks up all changes since T50.  Gets
    busy indexing the large number of updates.

  * Indexing finishes at T200

  * At T205, the indexer wakes up again and looks for updates since
    (T105 - 30 seconds).  That's the time it last looked for updates,
    minus a window to compensate for transactions that were still in
    flight when it last looked.

    Assuming (T105 - 30 seconds) is somewhere around T75, this means
    we once again see all of the updates that were committed at T100.
    We index them again!

  * People are sad

We've modified this to be a bit more selective.  The new behaviour:

  * When looking for updates, we separately check for any record with
    a timestamp between (LAST_CHECK - 30 seconds) and LAST_CHECK.

  * Some of these updates will be things we indexed last time, so
    subtract out that set of IDs to avoid handling them again.

  * Anything remaining is the result of a commit that was still in
    flight when we last checked, and needs to be indexed.

Since we only keep track of the IDs we indexed from the run
immediately prior, we can still end up doing some duplicate indexing
if the indexing rounds are running frequently.  But if the indexing
rounds are running frequently, they're probably not doing much anyway,
so this doesn't really hurt.

(cherry picked from commit f8dfb8ebc18c3411fc57dbc29e2b212c6bd47841)