QE25-76 Get rid of explicit boosting by record type
This history is a bit confusing.  It looks like the very first release
of ArchivesSpace was doing this kind of record type boosting, but it's
possible it never actually worked properly.

In any case, the config got fixed in 2019 and the boosting started
working again.  And it turns out we don't like it, so we're turning it
off.