Upgrade Sequel to version 5.95.1 to fix concurrency issue
Originally reported here:

  https://github.com/jeremyevans/sequel/issues/2292#issuecomment-2739067875

Soon after fixed here:

  https://github.com/jeremyevans/sequel/commit/9aaca73164454393d92036024f0383be9b42d7e9

We had originally applied a workaround, but it turns out not to be
fully effective because agent relationships themselves have a
relationship with subject terms.  Our "priming" of relationships only
went one level deep but would need to recurse to hit these second
order relationships.

Given that it's fixed up stream, it was more direct just to take the
fix and ditch the workaround.